Added back CloudFormation stack launch URLs.
Since we've switched over to using the Jam Platform for most deliveries of the workshop nowadays, we had removed the CloudFormation stack launch URL used in the past. This readds 2 URLs, one to launch in PDX/US Oregon/us-west-2, and one that will use the user's last used region (say that 5 times fast). This makes it easier for our customers/attendees to also use the workshop with their own AWS account if they choose to by simplifying the step of launching the CloudFormation stack.